+++ features/src/codegen/features/domain/ManyToManyAFooCodegen.java
@@ -17,9 +17,9 @@ import joist.util.Copy;
 public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
 
     public static final ManyToManyAFooQueries queries;
-    private Integer id = null;
+    private Long id = null;
     private String name = null;
-    private Integer version = null;
+    private Long version = null;
     private ForeignKeyListHolder<ManyToManyAFoo, ManyToManyAFooToBar> manyToManyAFooToBars = new ForeignKeyListHolder<ManyToManyAFoo, ManyToManyAFooToBar>((ManyToManyAFoo) this, Aliases.manyToManyAFooToBar(), Aliases.manyToManyAFooToBar().manyToManyAFoo);
     protected Changed changed;
 
@@ -37,11 +37,11 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
         this.addRule(new MaxLength<ManyToManyAFoo>(Shims.name, 100));
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.getChanged().record("id", this.id, id);
         this.id = id;
         if (UoW.isOpen()) {
@@ -58,7 +58,7 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
         this.name = name;
     }
 
-    public Integer getVersion() {
+    public Long getVersion() {
         return this.version;
     }
 
@@ -146,11 +146,11 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
     }
 
     static class Shims {
-        protected static final Shim<ManyToManyAFoo, Integer> id = new Shim<ManyToManyAFoo, Integer>() {
-            public void set(ManyToManyAFoo instance, Integer id) {
+        protected static final Shim<ManyToManyAFoo, Long> id = new Shim<ManyToManyAFoo, Long>() {
+            public void set(ManyToManyAFoo instance, Long id) {
                 ((ManyToManyAFooCodegen) instance).id = id;
             }
-            public Integer get(ManyToManyAFoo instance) {
+            public Long get(ManyToManyAFoo instance) {
                 return ((ManyToManyAFooCodegen) instance).id;
             }
             public String getName() {
@@ -168,11 +168,11 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
                 return "name";
             }
         };
-        protected static final Shim<ManyToManyAFoo, Integer> version = new Shim<ManyToManyAFoo, Integer>() {
-            public void set(ManyToManyAFoo instance, Integer version) {
+        protected static final Shim<ManyToManyAFoo, Long> version = new Shim<ManyToManyAFoo, Long>() {
+            public void set(ManyToManyAFoo instance, Long version) {
                 ((ManyToManyAFooCodegen) instance).version = version;
             }
-            public Integer get(ManyToManyAFoo instance) {
+            public Long get(ManyToManyAFoo instance) {
                 return ((ManyToManyAFooCodegen) instance).version;
             }
             public String getName() {
@@ -188,8 +188,8 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
         public boolean hasId() {
             return this.contains("id");
         }
-        public Integer getOriginalId() {
-            return (java.lang.Integer) this.getOriginal("id");
+        public Long getOriginalId() {
+            return (Long) this.getOriginal("id");
         }
         public boolean hasName() {
             return this.contains("name");
@@ -200,8 +200,8 @@ public abstract class ManyToManyAFooCodegen extends AbstractDomainObject {
         public boolean hasVersion() {
             return this.contains("version");
         }
-        public Integer getOriginalVersion() {
-            return (java.lang.Integer) this.getOriginal("version");
+        public Long getOriginalVersion() {
+            return (Long) this.getOriginal("version");
         }
         public boolean hasManyToManyAFooToBars() {
             return this.contains("manyToManyAFooToBars");
