+++ features/src/codegen/features/domain/ManyToManyBFooCodegen.java
@@ -17,9 +17,9 @@ import joist.util.Copy;
 public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
 
     public static final ManyToManyBFooQueries queries;
-    private Integer id = null;
+    private Long id = null;
     private String name = null;
-    private Integer version = null;
+    private Long version = null;
     private ForeignKeyListHolder<ManyToManyBFoo, ManyToManyBFooToBar> blueManyToManyBFooToBars = new ForeignKeyListHolder<ManyToManyBFoo, ManyToManyBFooToBar>((ManyToManyBFoo) this, Aliases.manyToManyBFooToBar(), Aliases.manyToManyBFooToBar().blue);
     protected Changed changed;
 
@@ -37,11 +37,11 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
         this.addRule(new MaxLength<ManyToManyBFoo>(Shims.name, 100));
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.getChanged().record("id", this.id, id);
         this.id = id;
         if (UoW.isOpen()) {
@@ -58,7 +58,7 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
         this.name = name;
     }
 
-    public Integer getVersion() {
+    public Long getVersion() {
         return this.version;
     }
 
@@ -146,11 +146,11 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
     }
 
     static class Shims {
-        protected static final Shim<ManyToManyBFoo, Integer> id = new Shim<ManyToManyBFoo, Integer>() {
-            public void set(ManyToManyBFoo instance, Integer id) {
+        protected static final Shim<ManyToManyBFoo, Long> id = new Shim<ManyToManyBFoo, Long>() {
+            public void set(ManyToManyBFoo instance, Long id) {
                 ((ManyToManyBFooCodegen) instance).id = id;
             }
-            public Integer get(ManyToManyBFoo instance) {
+            public Long get(ManyToManyBFoo instance) {
                 return ((ManyToManyBFooCodegen) instance).id;
             }
             public String getName() {
@@ -168,11 +168,11 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
                 return "name";
             }
         };
-        protected static final Shim<ManyToManyBFoo, Integer> version = new Shim<ManyToManyBFoo, Integer>() {
-            public void set(ManyToManyBFoo instance, Integer version) {
+        protected static final Shim<ManyToManyBFoo, Long> version = new Shim<ManyToManyBFoo, Long>() {
+            public void set(ManyToManyBFoo instance, Long version) {
                 ((ManyToManyBFooCodegen) instance).version = version;
             }
-            public Integer get(ManyToManyBFoo instance) {
+            public Long get(ManyToManyBFoo instance) {
                 return ((ManyToManyBFooCodegen) instance).version;
             }
             public String getName() {
@@ -188,8 +188,8 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
         public boolean hasId() {
             return this.contains("id");
         }
-        public Integer getOriginalId() {
-            return (java.lang.Integer) this.getOriginal("id");
+        public Long getOriginalId() {
+            return (Long) this.getOriginal("id");
         }
         public boolean hasName() {
             return this.contains("name");
@@ -200,8 +200,8 @@ public abstract class ManyToManyBFooCodegen extends AbstractDomainObject {
         public boolean hasVersion() {
             return this.contains("version");
         }
-        public Integer getOriginalVersion() {
-            return (java.lang.Integer) this.getOriginal("version");
+        public Long getOriginalVersion() {
+            return (Long) this.getOriginal("version");
         }
         public boolean hasBlueManyToManyBFooToBars() {
             return this.contains("blueManyToManyBFooToBars");
