+++ features/src/codegen/features/domain/builders/CodeADomainObjectBuilderCodegen.java
@@ -12,11 +12,11 @@ public abstract class CodeADomainObjectBuilderCodegen extends AbstractBuilder<Co
         super(instance);
     }
 
-    public Integer id() {
+    public Long id() {
         return get().getId();
     }
 
-    public CodeADomainObjectBuilder id(Integer id) {
+    public CodeADomainObjectBuilder id(Long id) {
         get().setId(id);
         return (CodeADomainObjectBuilder) this;
     }
